perf(app): 视图级懒加载代码分割与 Rust 启动路径瘦身
前端:
- 共享组件抽离:SettingsView 的 ComposerModelSelector/ScaledHtmlPreview/
  ComposerModeMenu/ComposerToolMenu 移入 composer-shared.jsx;Personas 的
  AppIcon/部门配色/personaText/Lanyard 移入 persona-shared.jsx——消除
  ChatView(启动视图)对两个巨型视图文件的静态牵连
- 8 个低频视图(Settings/Codex/卡池/工具商店/定时/知识库/监控/搜索)改
  React.lazy,VIEW_LOADERS 表统一出口;切视图包 startTransition(未就绪时
  保持旧视图不闪 fallback);NavItem 悬停/聚焦预取 + 首帧后空闲预取
  scheduled/settings(覆盖桌宠/定时快捷方式直跳入口)
- DetachedShell 整体 lazy,撕离窗与主窗共享同一批视图 chunk;主 chunk
  1129KB→352KB(gzip 303→102)
- 契约测试同步:7 个断言目标改为 lazy 工厂/composer-shared/persona-shared

Rust:
- lib.rs setup 删除三行重复 sidecar 加载(boot_inner 已加载)
- extraction.rs:15 处写盘改内容比对一致跳过(write_if_changed);mcp.json
  一次启动只读+parse 一遍(原两遍)且内容未变不写回;data_analysis 下架
  清理改残留探测,无残留不再每次启动实例化 MarketplaceManager+uninstall
- native_installer.rs:write_license 挪到 hash miss 后(已装好不重写);
  lark/wecom license 按目标平台 cfg 门控(非 Linux 构建不再嵌 Linux 文本)

验证: build:ui 通过;ui_smoke 47/47、settings 70、kb 53、toolstore 39、
composer 8、detached、markdown 10、acp-providers、update-notice、
pet-selector、knowledge-host 全绿;node --test 165/165;architecture-guard
通过;fork-guard --fast 通过;cargo fmt/check/test(runtime_bundle 17、
connectors 39)通过。scheduled_tasks_smoke 与 design_mode_entry_smoke 的
失败在 main 基线同样失败(预存 flake,与本次改动无关)。
