+++ pinvou3-app/tests/markdown_syntax_highlight.test.mjs
@@ -247,7 +247,7 @@ assert.match(
 // Personas.jsx 代码块明暗切换采用静态共存写法 light-code dark-code
 // （配合 base.css 的 .dark 祖先限定；P1 重构后 isDark 三元式已消除）。
 assert.match(
-  readApp('src', 'features', 'personas', 'Personas.jsx'),
+  readApp('src', 'features', 'personas', 'persona-shared.jsx'),
   /light-code dark-code/u,
 );
 console.log('Markdown syntax highlighting contract: ok');
